Split files for types/services
