+++ crates/conduit-wasm/src/lib.rs
@@ -2,11 +2,11 @@
 #![allow(clippy::large_enum_variant)]
 #![allow(clippy::eval_order_dependence)]
 
-pub mod agent;
 pub mod app;
 pub mod components;
 pub mod error;
 pub mod routes;
+pub mod services;
 pub mod types;
 
 use wasm_bindgen::prelude::*;
