+++ crates/conduit-wasm/src/routes/login.rs
@@ -5,10 +5,10 @@ use yew::{
 };
 use yew_router::{agent::RouteRequest::ChangeRoute, prelude::*};
 
-use crate::agent::{set_token, Auth};
 use crate::components::list_errors::ListErrors;
 use crate::error::Error;
 use crate::routes::AppRoute;
+use crate::services::{set_token, Auth};
 use crate::types::{LoginInfo, LoginInfoWrapper, UserInfo, UserInfoWrapper};
 
 /// Login page
