+++ crates/conduit-wasm/src/routes/register.rs
@@ -5,10 +5,10 @@ use yew::{
 };
 use yew_router::{agent::RouteRequest::ChangeRoute, prelude::*};
 
-use crate::agent::{set_token, Auth};
 use crate::components::list_errors::ListErrors;
 use crate::error::Error;
 use crate::routes::AppRoute;
+use crate::services::{set_token, Auth};
 use crate::types::{RegisterInfo, RegisterInfoWrapper, UserInfo, UserInfoWrapper};
 
 /// Register page
