--- crates/conduit-wasm/src/types.rs
@@ -1,172 +1,0 @@
-//! Common types
-
-use std::collections::HashMap;
-
-use chrono::prelude::*;
-use serde::{Deserialize, Serialize};
-
-/// Conduit api error info for Unprocessable Entity error
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ErrorInfo {
-    pub errors: HashMap<String, Vec<String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ArticleInfo {
-    pub slug: String,
-    pub title: String,
-    pub description: String,
-    pub body: String,
-    pub tag_list: Vec<String>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub favorited: bool,
-    pub favorites_count: u32,
-    pub author: ProfileInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ArticleInfoWrapper {
-    pub article: ArticleInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ArticleListInfo {
-    pub articles: Vec<ArticleInfo>,
-    pub articles_count: u32,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct ArticleCreateUpdateInfo {
-    pub title: String,
-    pub description: String,
-    pub body: String,
-    pub tag_list: Option<Vec<String>>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ArticleCreateUpdateInfoWrapper {
-    pub article: ArticleCreateUpdateInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct TagListInfo {
-    pub tags: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CommentInfo {
-    pub id: u32,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub body: String,
-    pub author: ProfileInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CommentInfoWrapper {
-    pub comment: CommentInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct CommentCreateInfo {
-    pub body: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CommentCreateInfoWrapper {
-    pub comment: CommentCreateInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CommentListInfo {
-    pub comments: Vec<CommentInfo>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct LoginInfo {
-    pub email: String,
-    pub password: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct LoginInfoWrapper {
-    pub user: LoginInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct RegisterInfo {
-    pub username: String,
-    pub email: String,
-    pub password: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct RegisterInfoWrapper {
-    pub user: RegisterInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct UserInfo {
-    pub email: String,
-    pub token: String,
-    pub username: String,
-    pub bio: Option<String>,
-    pub image: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct UserInfoWrapper {
-    pub user: UserInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct UserUpdateInfo {
-    pub email: String,
-    pub username: String,
-    pub password: Option<String>,
-    pub image: String,
-    pub bio: String,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct UserUpdateInfoWrapper {
-    pub user: UserUpdateInfo,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ProfileInfo {
-    pub username: String,
-    pub bio: Option<String>,
-    pub image: String,
-    pub following: bool,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ProfileInfoWrapper {
-    pub profile: ProfileInfo,
-}
-
-pub type DeleteWrapper = HashMap<(), ()>;
