+++ crates/conduit-wasm/src/types/articles.rs
@@ -1,0 +1,47 @@
+use chrono::prelude::*;
+use serde::{Deserialize, Serialize};
+
+use super::ProfileInfo;
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ArticleInfo {
+    pub slug: String,
+    pub title: String,
+    pub description: String,
+    pub body: String,
+    pub tag_list: Vec<String>,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub favorited: bool,
+    pub favorites_count: u32,
+    pub author: ProfileInfo,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ArticleInfoWrapper {
+    pub article: ArticleInfo,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ArticleListInfo {
+    pub articles: Vec<ArticleInfo>,
+    pub articles_count: u32,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct ArticleCreateUpdateInfo {
+    pub title: String,
+    pub description: String,
+    pub body: String,
+    pub tag_list: Option<Vec<String>>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ArticleCreateUpdateInfoWrapper {
+    pub article: ArticleCreateUpdateInfo,
+}
