+++ crates/conduit-wasm/src/types/comments.rs
@@ -1,0 +1,38 @@
+use chrono::prelude::*;
+use serde::{Deserialize, Serialize};
+
+use super::ProfileInfo;
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CommentInfo {
+    pub id: u32,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub body: String,
+    pub author: ProfileInfo,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CommentInfoWrapper {
+    pub comment: CommentInfo,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct CommentCreateInfo {
+    pub body: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CommentCreateInfoWrapper {
+    pub comment: CommentCreateInfo,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct CommentListInfo {
+    pub comments: Vec<CommentInfo>,
+}
