+++ crates/conduit-wasm/src/types/profiles.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ProfileInfo {
+    pub username: String,
+    pub bio: Option<String>,
+    pub image: String,
+    pub following: bool,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct ProfileInfoWrapper {
+    pub profile: ProfileInfo,
+}
