+++ crates/conduit-wasm/src/types/tags.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct TagListInfo {
+    pub tags: Vec<String>,
+}
